Another rebuild command fix
